app/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django==4.1.4
django-filter==22.1
et-xmlfile==1.1.0; python_version >= '3.6'
xlsxwriter==3.0.7
numpy==1.24.0; python_version >= '3.10'
openpyxl==3.0.10
pandas==1.5.2
# psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7
pyyaml==6.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3; python_version >= '3.5'
tzdata==2022.7; sys_platform == 'win32'
django-user-visit==1.1
django-crontab==0.7.1